+++ includes/auto-load-next-post-conditional-functions.php
@@ -38,14 +38,14 @@ if ( ! function_exists('supports_alnp')) {
 	 * Returns true or false if the plugin is supported by the theme.
 	 *
 	 * @since   1.4.3
-	 * @version 1.4.8
+	 * @version 1.4.9
 	 * @access  public
 	 * @return  boolen
 	 */
 	function supports_alnp() {
 		/* WordPress core themes. */
 		$core_themes = array(
-			'twentyseventeen', 'twentyfifteen', 'twentyfourteen', 'twentythirteen', 'twentytwelve', 'twentyten'
+			'twentyseventeen', 'twentysixteen', 'twentyfifteen', 'twentyfourteen', 'twentythirteen', 'twentytwelve', 'twentyten'
 		);
 
 		if (in_array(get_option('template'), $core_themes)) {
@@ -63,6 +63,7 @@ if ( ! function_exists('alnp_template_location')) {
 	 * Filters the template location for get_template_part().
 	 *
 	 * @since   1.4.8
+	 * @version 1.4.9
 	 * @access  public
 	 * @return  boolen
 	 */
@@ -72,6 +73,10 @@ if ( ! function_exists('alnp_template_location')) {
 		switch( $current_theme ) {
 			case 'twentyseventeen':
 				$path = 'template-parts/post/';
+				break;
+
+			case 'twentysixteen':
+				$path = 'template-parts/';
 				break;
 
 			default: